Uses schema_io for optional property definition

+++ samples/client/petstore/python/src/petstore_api/components/schema/object_with_non_intersecting_values.py
@@ -79,8 +79,8 @@ ObjectWithNonIntersectingValuesDictInput = typing.Mapping[
     str,
     typing.Union[
         typing.Union[
-            float,
-            int
+            int,
+            float
         ],
         str,
     ]
