+++ samples/client/petstore/python/src/petstore_api/components/schema/object_with_only_optional_props.py
@@ -86,8 +86,8 @@ ObjectWithOnlyOptionalPropsDictInput = typing_extensions.TypedDict(
     {
         "a": str,
         "b": typing.Union[
-            float,
-            int
+            int,
+            float
         ],
     },
     total=False
